sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mnemo-client"
version = "0.3.1"
description = "Python client for Mnemo memory API"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Mnemo" }]
license = { text = "Apache-2.0" }
dependencies = []

[project.optional-dependencies]
async = ["aiohttp>=3.9"]
langchain = ["langchain-core>=0.2"]
llamaindex = ["llama-index-core>=0.10"]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "requests>=2.31"]
all = ["mnemo-client[async,langchain,llamaindex]"]

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["mnemo*"]
